feat(components/Dialog): updated tests
