+++ packages/components/src/Dialog/Dialog.test.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import classNames from 'classnames';
+import { shallow } from 'enzyme';
 import Dialog from './Dialog.component';
 
 function mockFakeComponent(name) {
@@ -15,7 +16,6 @@ function mockFakeComponent(name) {
 	return fakeComponent;
 }
 
-jest.mock('react-dom');
 jest.mock('react-bootstrap/lib/Modal', () => {
 	const Modal = mockFakeComponent('Modal');
 	Modal.Header = mockFakeComponent('Header');
@@ -72,21 +72,6 @@ const largeProps = {
 	},
 };
 
-const headerPropsNullDialogProps = {
-	header: 'Hello world',
-	bsDialogProps: null,
-	show: true,
-};
-const headerPropsUndefinedDialogProps = {
-	header: 'Hello world',
-	bsDialogProps: undefined,
-	show: true,
-};
-const headerNoDialogPropsProps = {
-	header: 'Hello world',
-	show: true,
-};
-
 const children = (<div>BODY</div>);
 
 describe('Dialog', () => {
@@ -119,25 +104,5 @@ describe('Dialog', () => {
 			<Dialog {...largeProps}>{children}</Dialog>
 		).toJSON();
 		expect(wrapper).toMatchSnapshot();
-	});
-	it('should render dialog without bsDialogProps', () => {
-		const customProps = { ...headerProps, undefined };
-		expect(customProps.header).toEqual('Hello world');
-		expect(customProps.show).toEqual(true);
-
-		const wrapperNullDialogProps = renderer.create(
-			<Dialog {...headerPropsNullDialogProps}>{children}</Dialog>
-		).toJSON();
-		expect(wrapperNullDialogProps).toMatchSnapshot();
-
-		const wrapperUndefinedDialogProps = renderer.create(
-			<Dialog {...headerPropsUndefinedDialogProps}>{children}</Dialog>
-		).toJSON();
-		expect(wrapperUndefinedDialogProps).toMatchSnapshot();
-
-		const wrapperNoDialogPropsProps = renderer.create(
-			<Dialog {...headerNoDialogPropsProps}>{children}</Dialog>
-		).toJSON();
-		expect(wrapperNoDialogPropsProps).toMatchSnapshot();
 	});
 });
